serverless/lambda_function_worker/requirements.txt
boto3
Pillow
REQ